port/mtx: C_MTXIdentity never wrote the translation column
The off-world second pass, from M4 §14.2, is one missing line in the SDK's
own C matrix library.

`C_MTXIdentity` sets the 3x3 and stops.  On the console that is invisible:
`MTXIdentity` is `PSMTXIdentity`, six paired-single stores that cover all
twelve elements, so the translation column is always zeroed and no caller
ever had to zero it itself.  `-DMTX_USE_C` makes the C body the one that
runs, and `src/game/hsfdraw.c:mtxRot` calls it on an *uninitialised stack*
`Mtx` whenever the x rotation is zero:

    if (x != 0.0f) MTXRotRad(mtx, 'X', ...); else MTXIdentity(mtx);
    ... MTXConcat(rotY, mtx, mtx) ... mtxTransCat(mtx, pos.x, pos.y, pos.z);

so the matrix keeps whatever translation was in that stack slot, and
`mtxTransCat` adds the model's position to it.  The slot is the same one
every call, so the value accumulates: `Hu3DPostExec`'s `sp40` and
`Hu3DShadowExec`'s `sp58` walk to tens of thousands over a few hundred
models.  Every object whose x rotation happens to be non-zero takes the
`MTXRotRad` branch, which does write all twelve, and is drawn correctly --
which is exactly the pattern M4 measured: the stage (rot.x == 0) always
off-world, the characters drawn twice because the shadow pass and the main
pass disagree about which branch they took, and the camera, the model
placements and the envelope matrices all provably sane, because none of them
goes through `mtxRot`.

--scenelog's envelope line closes the last of M4's three candidates: at frame
5100 every `hsf->matrix->data[]` translation is under 439.  The other two,
objReplica/objMap and the hookMdlId chain, are downstream of the same
`mtxRot`.

Fixed in patches.txt, because it is the decomp's file and the decomp is
right about the DOL.  `uintptr_t`-style: the change is invisible to the
console build.

Proved rather than asserted:

  * port/tests/mtx_test.c, a new host test, poisons the destination and
    demands every constructor write all twelve elements, runs every
    in-place form against the out-of-place answer, diffs every PS*
    replacement against the C_* body it stands in for, and round-trips
    PSMTXReorder + PSMTXROMultVecArray against C_MTXMultVecArray.  It found
    this in one run and reports nothing else wrong.  `make -C port
    TARGET=host mtxtest`.
  * `port> GX draw: N primitive(s) off-world` is now a permanent counter
    (one compare per primitive).  Over the 5,150-frame menu walk on the G4:
    **0**, against M4's 1,264 of 1,756 on one frame.  Primitive and vertex
    counts are unchanged to the digit -- the same geometry, different
    matrices.
  * docs/screenshots/mp4-charselect-stage.png: the theatre stage is there.

Also in here, and unrelated except in kind: ARAM transfers carry main-memory
addresses in `u32` SDK fields, and `ArqCallBackAM` casts one straight back to
a pointer.  Every cast on the game's side was already widened by
widen_ptr_casts.py; the truncation left was in the SDK's own prototypes.
Widened in patches.txt -- identity on the G4, and the difference between a
segfault in `CharInit` and reaching the title screen on a 64-bit host.

+++ port/src/audio/aram.c
@@ -5,6 +5,11 @@
  * allocator over plain `u32` offsets (`AMEM_PTR`) above HU_AMEM_BASE, and every
  * transfer goes through `ARQPostRequest`, which becomes a memcpy.
  *
+ * The one thing that is not free on a 64-bit host is that a `u32` in this
+ * interface is sometimes an ARAM offset and sometimes a main-memory address;
+ * patches.txt widens the latter to `uintptr_t` in the SDK's own headers, which
+ * is the identity on the G4.
+ *
  * MusyX owns everything below HU_AMEM_BASE (8.03 MB of sample data) and the
  * game the 8.39 MB above it, and neither cares that the "DMA" completed before
  * the call returned -- armem.c counts outstanding requests in `arqCnt` and
@@ -48,15 +53,15 @@ ARCallback ARRegisterDMACallback(ARCallback cb) {
     return old;
 }
 
-void ARStartDMA(u32 type, u32 mainmem_addr, u32 aram_addr, u32 length) {
+void ARStartDMA(u32 type, uintptr_t mainmem_addr, u32 aram_addr, u32 length) {
     u8* aram = (u8*)port_aram();
     if (aram_addr + length > PORT_ARAM_SIZE) {
         port_fatal("ARStartDMA out of range: %08x + %u", aram_addr, length);
     }
     if (type == ARAM_DIR_MRAM_TO_ARAM) {
-        memcpy(aram + aram_addr, (void*)(uintptr_t)mainmem_addr, length);
+        memcpy(aram + aram_addr, (void*)mainmem_addr, length);
     } else {
-        memcpy((void*)(uintptr_t)mainmem_addr, aram + aram_addr, length);
+        memcpy((void*)mainmem_addr, aram + aram_addr, length);
     }
     dma_bytes += length;
     if (dma_callback) {
@@ -80,8 +85,8 @@ void ARQRemoveRequest(ARQRequest* task) { (void)task; }
 void ARQRemoveOwnerRequest(u32 owner) { (void)owner; }
 void ARQFlushQueue(void) {}
 
-void ARQPostRequest(ARQRequest* task, u32 owner, u32 type, u32 priority, u32 source, u32 dest,
-                    u32 length, ARQCallback callback) {
+void ARQPostRequest(ARQRequest* task, u32 owner, u32 type, u32 priority, uintptr_t source,
+                    uintptr_t dest, u32 length, ARQCallback callback) {
     task->next = NULL;
     task->owner = owner;
     task->type = type;
@@ -96,7 +101,7 @@ void ARQPostRequest(ARQRequest* task, u32 owner, u32 type, u32 priority, u32 sou
         ARStartDMA(ARAM_DIR_ARAM_TO_MRAM, dest, source, length);
     }
     if (callback) {
-        callback((u32)(uintptr_t)task);
+        callback((uintptr_t)task);
     }
 }
